+++ src/hooks/ready/checkCompendiums.js
@@ -39,6 +39,7 @@ export default async function () {
       createIfNotExists("entity-feature-compendium", "Item", "Class Features"),
       createIfNotExists("entity-class-compendium", "Item", "Classes"),
       createIfNotExists("entity-trait-compendium", "Item", "Racial Traits"),
+      createIfNotExists("entity-feat-compendium", "Item", "Feats"),
       createIfNotExists("entity-race-compendium", "Item", "Races"),
       createIfNotExists("entity-monster-compendium", "Actor", "Monsters"),
       // createIfNotExists("entity-monster-feature-compendium", "Item", "Monster Features")