+++ src/muncher/feats.js
@@ -1,0 +1,44 @@
+// Main module class
+import { munchNote, getCampaignId, download } from "./utils.js";
+import { getFeats } from "./feats/feats.js";
+
+function getFeatData() {
+  const cobaltCookie = game.settings.get("ddb-importer", "cobalt-cookie");
+  const campaignId = getCampaignId();
+  const parsingApi = game.settings.get("ddb-importer", "api-endpoint");
+  const betaKey = game.settings.get("ddb-importer", "beta-key");
+  const body = { cobalt: cobaltCookie, campaignId: campaignId, betaKey: betaKey };
+  const debugJson = game.settings.get("ddb-importer", "debug-json");
+
+  return new Promise((resolve, reject) => {
+    fetch(`${parsingApi}/proxy/feats`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(body),
+    })
+      .then((response) => response.json())
+      .then((data) => {
+        if (debugJson) {
+          download(JSON.stringify(data), `feats-raw.json`, "application/json");
+        }
+        if (!data.success) {
+          munchNote(`Failure: ${data.message}`);
+          reject(data.message);
+        }
+        return data;
+      })
+      .then((data) => getFeats(data.data))
+      .then((data) => resolve(data))
+      .catch((error) => reject(error));
+  });
+}
+
+export async function parseFeats() {
+  const results = await getFeatData();
+
+  return results;
+}
+
+