+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * Munching Spells and Items can now copy over DAE SRD effects.
 * Support Silvered, Magical and Adamantine weapon properties.
 * Retain journal notes on characters.
-* Some NPC's like Nezznar did not match some spells due to formating differences.
+* Some NPC's like Nezznar did not match some spells due to formatting differences.
+* Another bug squashed preventing npcs importing from SRD compendiums.
 
 # 0.2.21